+++ fabric/src/main/java/io/github/mortuusars/exposure/fabric/ExposureFabric.java
@@ -10,6 +10,7 @@ import io.github.mortuusars.exposure.command.TestCommand;
 import io.github.mortuusars.exposure.data.Lenses;
 import io.github.mortuusars.exposure.fabric.integration.create.CreateFilmDeveloping;
 import io.github.mortuusars.exposure.fabric.resources.FabricLensesDataLoader;
+import io.github.mortuusars.exposure.integration.ModCompatibility;
 import io.github.mortuusars.exposure.network.fabric.PacketsImpl;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
@@ -39,6 +40,8 @@ public class ExposureFabric implements ModInitializer {
             if (config.getType() == ModConfig.Type.COMMON && FabricLoader.getInstance().isModLoaded("create")) {
                 CreateFilmDeveloping.clearCachedData();
             }
+
+            ModCompatibility.handle();
         });
 
         ForgeConfigRegistry.INSTANCE.register(Exposure.ID, ModConfig.Type.COMMON, Config.Common.SPEC);