+++ forge/src/main/java/io/github/mortuusars/exposure/forge/event/CommonEvents.java
@@ -6,6 +6,7 @@ import io.github.mortuusars.exposure.command.ShaderCommand;
 import io.github.mortuusars.exposure.command.TestCommand;
 import io.github.mortuusars.exposure.data.Lenses;
 import io.github.mortuusars.exposure.data.LensesDataLoader;
+import io.github.mortuusars.exposure.integration.ModCompatibility;
 import io.github.mortuusars.exposure.network.forge.PacketsImpl;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraftforge.event.AddReloadListenerEvent;
@@ -24,6 +25,7 @@ public class CommonEvents {
                 PacketsImpl.register();
                 Exposure.Advancements.register();
                 Exposure.Stats.register();
+                ModCompatibility.handle();
             });
         }